upload changes for workers
